pyproject.toml
[project]
name = "metadata-governance-toolkit"
version = "0.1.0"
description = "Metadata scanning + data dictionary + MetaDB + API + scheduler"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "SQLAlchemy>=2.0",
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "python-dotenv>=1.0",
  "APScheduler>=3.10",
  "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
  "ruff>=0.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
